+++ packages/visual-editor/src/editor/YextField.tsx
@@ -23,7 +23,6 @@ import {
 import { RenderEntityFieldFilter } from "../internal/utils/getFilteredEntityFields.ts";
 import { MsgString } from "../utils/i18n/platform.ts";
 import { IMAGE_CONSTANT_CONFIG } from "../internal/puck/constant-value-fields/Image.tsx";
-import { VIDEO_CONSTANT_CONFIG } from "../internal/puck/constant-value-fields/Video.tsx";
 import type { YextPuckFields } from "../fields/fields.ts";
 
 /** Copied from Puck, do not change */
@@ -106,10 +105,6 @@ type YextImageField = YextBaseField & {
   type: "image";
 };
 
-type YextVideoField = YextBaseField & {
-  type: "video";
-};
-
 type YextDynamicSelectField<T extends DynamicOptionValueTypes> =
   YextBaseField & {
     type: "dynamicSelect";
@@ -134,11 +129,10 @@ type YextFieldConfig<Props = any> =
   | YextMaxWidthField
   | YextTranslatableStringField
   | YextImageField
-  | YextVideoField
   | YextDynamicSelectField<Props extends DynamicOptionValueTypes ? Props : any>
   | YextPuckFields[Exclude<
       keyof YextPuckFields,
-      "basicSelector" | "optionalNumber"
+      "basicSelector" | "optionalNumber" | "video"
     >];
 
 export function YextField<T = any>(
@@ -211,13 +205,6 @@ export function YextField<T, U>(
   if (config.type === "image") {
     return {
       ...IMAGE_CONSTANT_CONFIG,
-      label: fieldName,
-    };
-  }
-
-  if (config.type === "video") {
-    return {
-      ...VIDEO_CONSTANT_CONFIG,
       label: fieldName,
     };
   }
