+++ packages/visual-editor/src/fields/index.ts
@@ -25,3 +25,5 @@ export {
   OptionalNumberFieldOverride,
   type OptionalNumberField,
 } from "./OptionalNumberField.tsx";
+
+export { VideoFieldOverride, type VideoField } from "./VideoField.tsx";
